pyproject.toml
[project]
name = "agentic-system-patternss"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.57.1",
    "google-genai>=1.25.0",
    "gradio>=5.37.0",
    "ipython>=9.4.0",
    "openai>=1.96.0",
    "openai-agents>=0.2.0",
    "pydantic>=2.11.7",
    "pypdf>=5.8.0",
    "python-dotenv>=1.1.1",
]
